backend/deprecated/requirements/local.txt
-r ./base.txt

# to read pyproject.toml configuration
toml==0.10.1

# better terminal
ipython==7.14.0

# for unit tests
factory-boy==2.12.0
pytest-django==3.9.0
pytest-mock==3.1.0

# for formatting and linters
isort==4.3.21
black==19.10b0
flake8==3.8.1
flake8-bugbear==20.1.4
